Added abstract FromToModel that add from to fields and does the validation in issue #53
